+++ cyxwiz-engine/src/gui/loaders/data_loader.h
@@ -80,8 +80,8 @@ struct AsyncLoadState {
     std::shared_ptr<cyxwiz::ParquetBackedDataset> previous_parquet_dataset;
     std::string previous_source_path;
     std::string source_path;         // file_path or folder_path that was loaded
-    int num_classes = 0;             // set by image/audio/text loaders
-    int vocab_size = 0;              // set by text loader
+    size_t num_classes = 0;          // set by image/audio/text loaders
+    size_t vocab_size = 0;           // set by text loader
     int audit_errors = 0;            // metadata/data sanity audit result
     int audit_warnings = 0;
     std::string audit_message;       // compact status suffix for the UI
