+++ cyxwiz-engine/src/gui/panels/differentiation_panel.cpp
@@ -184,7 +184,8 @@ void DifferentiationPanel::RenderMethodSelector() {
 
 void DifferentiationPanel::RenderLoadingIndicator() {
     ImGui::Text("Computing derivatives...");
-    float progress = 0.5f + 0.5f * std::sin(ImGui::GetTime() * 5.0f);
+    float progress = 0.5f + 0.5f * static_cast<float>(
+        std::sin(ImGui::GetTime() * 5.0));
     ImGui::ProgressBar(progress, ImVec2(-1, 0));
 }
