+++ cyxwiz-engine/src/gui/panels/qp_panel.cpp
@@ -225,7 +225,8 @@ void QPPanel::RenderConstraints() {
 
 void QPPanel::RenderLoadingIndicator() {
     ImGui::Text("Solving QP problem...");
-    float progress = 0.5f + 0.5f * std::sin(ImGui::GetTime() * 5.0f);
+    float progress = 0.5f + 0.5f * static_cast<float>(
+        std::sin(ImGui::GetTime() * 5.0));
     ImGui::ProgressBar(progress, ImVec2(-1, 0));
 }
